backend/requirements.txt
fastapi
uvicorn
httpx
sqlalchemy
pydantic
pydantic-settings
python-dotenv
openai
